Time: 0 ms (100.00%) | Memory: 8.5 MB (88.37%) - LeetSync

+++ 1584-average-salary-excluding-the-minimum-and-maximum-salary/average-salary-excluding-the-minimum-and-maximum-salary.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
     double average(vector<int>& salary) {
         sort(salary.begin(), salary.end());
-        double n=salary.size();
+        int n=salary.size();
         double sum =0;
         for(int i=1; i<n-1; i++){
             sum += salary[i];
         };
-        double count = n-2;
+        int count = n-2;
         return sum/count;
     }
 };
